Move import into function
