+++ executorlib/standalone/hdf.py
@@ -1,4 +1,3 @@
-import os
 from typing import Any, Optional
 
 import cloudpickle
@@ -98,25 +97,17 @@ def get_runtime(file_name: str) -> float:
 
 
 def get_queue_id(file_name: Optional[str]) -> Optional[int]:
+    """
+    Get queuing system id from HDF5 file
+
+    Args:
+        file_name (str): file name of the HDF5 file as absolute path
+
+    Returns:
+        int: queuing system id from the execution of the python function
+    """
     if file_name is not None:
         with h5py.File(file_name, "r") as hdf:
             if "queue_id" in hdf:
                 return cloudpickle.loads(np.void(hdf["/queue_id"]))
     return None
-
-
-def get_cache_data(cache_directory: str) -> list[dict]:
-    file_lst = []
-    for task_key in os.listdir(cache_directory):
-        file_name = os.path.join(cache_directory, task_key, "cache.h5out")
-        os.makedirs(os.path.join(cache_directory, task_key), exist_ok=True)
-        if os.path.exists(file_name):
-            with h5py.File(file_name, "r") as hdf:
-                file_content_dict = {
-                    key: cloudpickle.loads(np.void(hdf["/" + key]))
-                    for key in group_dict.values()
-                    if key in hdf
-                }
-            file_content_dict["filename"] = file_name
-            file_lst.append(file_content_dict)
-    return file_lst
